chore: fix

+++ Test/Passes/InstructionSelection/RISCV64/li.mlir
@@ -2,9 +2,9 @@
 
 "builtin.module"() ({
     "func.func"() ({
-        %one = "llvm.constant"() <{ "value" = 1 : i64 }> : () -> i64
-        %two = "llvm.constant"() <{ "value" = 2 : i64 }> : () -> i64
-        // CHECK:      %{{.*}} = "riscv.li"() <{"value" = 1 : i64}> : () -> i64
-        // CHECK-NEXT: %{{.*}} = "riscv.li"() <{"value" = 2 : i64}> : () -> i64
+        %one = "llvm.constant"() <{ "value" = 1 : i64 }> : () -> i8
+        %two = "llvm.constant"() <{ "value" = 2 : i64 }> : () -> i8
+        // CHECK:      %{{.*}} = "riscv.li"() <{"value" = 1 : i64}> : () -> !reg
+        // CHECK-NEXT: %{{.*}} = "riscv.li"() <{"value" = 2 : i64}> : () -> !reg
     }) : () -> ()
 }) : () -> ()
